+++ SegmentTree/SegmentTree/CSegmentTree.h
@@ -6,45 +6,45 @@
 
 struct SegmentTreeData
 {
-	long root = 1;
-	long left{};
-	long right{};
-	long middle{};
+	long long root = 1;
+	long long left{};
+	long long right{};
+	long long middle{};
 };
 
 class CSegmentTree
 {
 public:
-	CSegmentTree(std::istream& input, std::ostream& output, SegmentTreeData& treeData, const std::vector<int>& arr);
+	CSegmentTree(std::istream& input, std::ostream& output, SegmentTreeData& treeData, const std::vector<long long>& arr);
 
 	void BuildTree();
 	void HandleCommand();
 	void PrintInfo();
 
 private:
-	void Build(long root, long left, long right);
+	void Build(long long root, long long left, long long right);
 
-	void Add(long root, long tLeft, long tRight, long index, long value);
-	void AddInt(long root, long tLeft, long tRight, long left, long right, long value);
+	void Add(long long root, long long tLeft, long long tRight, long long index, long long value);
+	void AddInt(long long root, long long tLeft, long long tRight, long long left, long long right, long long value);
 
-	long Rsq(long root, long tLeft, long tRight, long left, long right);
+	long long Rsq(long long root, long long tLeft, long long tRight, long long left, long long right);
 
-	std::vector<long> GetNumbersFromStream(std::istringstream& iss);
+	std::vector<long long> GetNumbersFromStream(std::istringstream& iss);
 
-	SegmentTreeData GetSegmentTreeData(long root, long left, long right);
-	long GetLeftIndex(long root);
-	long GetRightIndex(long root);
+	SegmentTreeData GetSegmentTreeData(long long root, long long left, long long right);
+	long long GetLeftIndex(long long root);
+	long long GetRightIndex(long long root);
 
 private:
 	std::istream& m_input;
 	std::ostream& m_output;
 
 	SegmentTreeData& m_treeData;
-	std::vector<long> m_tree;
-	std::vector<long> m_add;
+	std::vector<long long> m_tree;
+	std::vector<long long> m_add;
 
 	std::vector<std::string> m_logger;
 
-	const std::vector<int> m_array;
+	const std::vector<long long> m_array;
 };
